backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
supabase==2.11.0
httpx==0.27.2
websockets==13.1
pandas==2.3.3
bcrypt==4.1.2
pyjwt==2.8.0
gunicorn==21.2.0

xgboost==3.1.3
psutil==7.2.1

numpy==2.2.6
mlflow==3.8.1

# Test dependencies for FlavorFit

# Core testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Web testing
selenium==4.15.2
webdriver-manager==4.0.1

# Code quality
flake8==6.1.0
black==23.11.0
pylint==3.0.2

# Type checking
mypy==1.7.1

# Coverage
coverage[toml]==7.3.2
